+++ Model/Group.php
@@ -11,8 +11,12 @@
 
 namespace FOS\UserBundle\Model;
 
+@trigger_error('Using Groups is deprecated since version 2.2 and will be removed in 3.0.', E_USER_DEPRECATED);
+
 /**
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ *
+ * @deprecated
  */
 abstract class Group implements GroupInterface
 {